collapse module cleanup (adjacency/overlapping separated)
