+++ grid_forge/src/gen/collapse/adjacency/frequency.rs
