+++ grid_forge/src/gen/collapse/adjacency/rules.rs
