--- grid_forge/src/gen/collapse/rules/mod.rs
@@ -1,5 +1,0 @@
-mod adjacency;
-mod pattern;
-
-pub use adjacency::*;
-// pub use pattern::*;
